# docs/dev/api/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  description: 'Authentication API: accounts, sessions, second factors and client applications. Issues ES256 access tokens and rotating refresh tokens, and publishes the JWKS resource servers verify them with.'
  license:
    name: MIT
  version: 2.0.1
servers:
- url: http://localhost:3000
  description: Local development
paths:
...
  /users/me/sessions:
    get:
      tags:
      - sessions
      operationId: list
      responses:
        '200':
          description: Active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
    delete:
      tags:
      - sessions
      operationId: revoke_all
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: 'null'
              - $ref: '#/components/schemas/RevokeAllRequest'
      responses:
        '204':
          description: Every other session revoked, and the current one too unless `keep_current_session`
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/sessions/{id}:
    delete:
      tags:
      - sessions
      operationId: revoke
      parameters:
      - name: id
        in: path
        description: Method or session id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: 'null'
              - $ref: '#/components/schemas/RevokeAllRequest'
      responses:
        '204':
          description: Session revoked
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
components:
  schemas:
...
    ErrorBody:
      type: object
      description: Body of every error response.
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 'Stable and machine-readable: `invalid_credentials`, `reauthentication_required`, ...'
        message:
          type: string
          description: Human-readable explanation; may change between versions.
...
    RevokeAllRequest:
      type: object
      properties:
        current_password:
          type:
          - string
          - 'null'
        keep_current_session:
          type: boolean
          description: 'Keep the session making the request signed in. Default: false.'
...
    SessionResponse:
      type: object
      required:
      - id
      - session_type
      - last_used_at
      - expires_at
      - created_at
      - is_current
      properties:
        client_id:
          type:
          - string
          - 'null'
        created_at:
          type: integer
          format: int64
        device_name:
          type:
          - string
          - 'null'
        expires_at:
          type: integer
          format: int64
        id:
          type: string
          format: uuid
        ip_address:
          type:
          - string
          - 'null'
        is_current:
          type: boolean
          description: True when this is the session used to make the current request.
        last_used_at:
          type: integer
          format: int64
        session_type:
          $ref: '#/components/schemas/SessionType'
        user_agent:
          type:
          - string
          - 'null'
    SessionType:
      type: string
      enum:
      - web
      - device
      - personal_access_token
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...